+++ packages/react/runtime/src/lynx.ts
@@ -5,6 +5,8 @@ import { options } from 'preact';
 // to make sure preact's hooks to register earlier than ours
 import './hooks/react.js';
 
+import { initAlog } from './alog/index.js';
+import { setupComponentStack } from './debug/component-stack.js';
 import { initProfileHook } from './debug/profile.js';
 import { document, setupBackgroundDocument } from './document.js';
 import { replaceCommitHook } from './lifecycle/patch/commit.js';
@@ -34,10 +36,19 @@ if (__MAIN_THREAD__) {
   }
 }
 
+if (__DEV__) {
+  setupComponentStack();
+}
+
 // TODO: replace this with __PROFILE__
 if (__PROFILE__) {
   // We are profiling both main-thread and background.
   initProfileHook();
+}
+
+if (typeof __ALOG__ !== 'undefined' && __ALOG__) {
+  // We are logging both main-thread and background.
+  initAlog();
 }
 
 if (__BACKGROUND__) {