+++ packages/react/runtime/vitest.config.ts
@@ -74,6 +74,7 @@ export default defineConfig({
     name: 'react/runtime',
     coverage: {
       exclude: [
+        'debug',
         'jsx-runtime',
         'jsx-dev-runtime',
         'lepus/jsx-dev-runtime',
@@ -85,6 +86,7 @@ export default defineConfig({
         'src/lynx-api.ts',
         'src/lynx.ts',
         'src/root.ts',
+        'src/debug/component-stack.ts',
         'src/debug/debug.ts',
         'src/lynx/calledByNative.ts',
         'src/lynx/component.ts',