+++ packages/third-party/tailwind-preset/src/plugins/lynx/index.ts
@@ -25,3 +25,7 @@ export { wordBreak } from './wordBreak.js';
 export { gridColumn } from './gridColumn.js';
 export { gridRow } from './gridRow.js';
 export { perspective } from './perspective.js';
+export { soloTranslate } from './soloTranslate.js';
+export { soloRotate } from './soloRotate.js';
+export { soloScale } from './soloScale.js';
+export { soloSkew } from './soloSkew.js';