+++ packages/web-platform/inline-style-parser/src/tokenize.rs
@@ -1,5 +1,5 @@
-use crate::parser::{char_code_definitions, types::*, utils::*};
 use crate::*;
+use crate::{char_code_definitions, types::*, utils::*};
 
 const URL_STR: [u16; 3] = ['u' as u16, 'r' as u16, 'l' as u16];
 /**